+++ spec/features/evidence_feature_spec.rb
@@ -27,22 +27,7 @@ end
 
 describe "viewing evidence" do
   it "should allow the user to view evidence" do
-    create_achievement
-    expect(page).not_to have_content "No achievements yet."
-    click_link "Attended career fair"
-    expect(page).to have_content "Attended career fair"
-    expect(page).to have_content "Attended UCL career fair to represent HSBC IT. Spoke with a load of grads."
-    expect(page).to have_content "Sponsorship, Role Model"
-    expect(page).to have_content  "2017-01-21"
+  create_endorsement
+  expect(page).to have_content " "
   end
-end
-
-def create_achievement
-  visit "/evidences"
-  click_link("Add achievement")
-  fill_in "Headline", with: "Attended career fair"
-  fill_in "Description", with: "Attended UCL career fair to represent HSBC IT. Spoke with a load of grads."
-  fill_in "Category", with: "Sponsorship, Role Model"
-  fill_in "Date achieved", with: "2017-01-21"
-  click_button "Submit"
 end
